Perk work
Implemented the UI for editing Perks and their entry points. Made some minor fixes to the backend, and improved how we represent entry points within DovahKit (though further improvements are needed there).

Made some UB fixes to some of the self-sorting UI models. In the next commit, I plan to make more use of a helper function that I already had and am using in the model for perks' entries.

+++ native/src/DovahKit/dovah/forms/structs/perk_entry_point_data.h
@@ -34,7 +34,7 @@ namespace dovah::loaded_forms::structs {
                public:
                   struct flag {
                      flag() = delete;
-                     enum type : uint8_t {
+                     enum type : uint16_t {
                         run_immediately = 1 << 0,
                         replace_default = 1 << 1,
                      };
@@ -44,7 +44,7 @@ namespace dovah::loaded_forms::structs {
                public:
                   form_reference_t spell;
                   localized_string text;
-                  uint8_t          flags = 0;
+                  uint16_t         flags = 0;
                   uint16_t         fragment_index = no_fragment;
             };
             struct spell {
